project controller fix error codes

+++ backend/todo-api/src/main/java/com/naatho/todoapp/contoller/ProjectController.java
@@ -1,8 +1,10 @@
 package com.naatho.todoapp.contoller;
 
+import com.naatho.todoapp.entity.Label;
 import com.naatho.todoapp.entity.Project;
 import com.naatho.todoapp.service.ProjectService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,9 +21,9 @@ public class ProjectController {
 
     @PostMapping()
     public @ResponseBody
-    String addNewProject (@Valid @RequestBody Project project) {
+    ResponseEntity<Label> addNewProject (@Valid @RequestBody Project project) {
         projectService.save(project);
-        return "Saved";
+        return ResponseEntity.ok().build();
     }
 
     @GetMapping()
@@ -31,13 +33,26 @@ public class ProjectController {
 
     @GetMapping(path="/{id}")
     public @ResponseBody
-    Optional<Project> getProjectByID(@PathVariable Integer id) {
-        return projectService.findById(id);
+    ResponseEntity<Project> getProjectByID(@PathVariable Integer id) {
+        Optional<Project> foundProject = projectService.findById(id);
+
+        if (foundProject.isEmpty()) {
+            return ResponseEntity.notFound().build();
+
+        } else {
+            return ResponseEntity.ok(foundProject.get());
+        }
     }
 
     @DeleteMapping(path="/{id}")
     public @ResponseBody
-    void deleteProjectById(@PathVariable Integer id) {
-        projectService.deleteById(id);
+    ResponseEntity<Project> deleteProjectById(@PathVariable Integer id) {
+        Optional<Project> foundProject = projectService.findById(id);
+        if (foundProject.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        } else {
+            projectService.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
     }
 }
